@@ -1,9 +1,8 @@
 {
   "name": "hoist-non-react-statics",
-  "version": "2.5.5",
+  "version": "3.3.0",
   "description": "Copies non-react specific statics from a child component to a parent component",
   "main": "dist/hoist-non-react-statics.cjs.js",
-  "types": "index.d.ts",
   "repository": {
     "type": "git",
     "url": "git://github.com/mridgway/hoist-non-react-statics.git"
@@ -14,7 +13,7 @@
     "index.d.ts"
   ],
   "scripts": {
-    "cover": "node node_modules/istanbul/lib/cli.js cover --dir artifacts -- ./node_modules/mocha/bin/_mocha tests/unit/ --recursive --compilers js:babel/register --reporter spec",
+    "cover": "node node_modules/istanbul/lib/cli.js cover --dir artifacts -- ./node_modules/mocha/bin/_mocha tests/unit/ --recursive --require babel/register --reporter spec",
     "lint": "eslint src",
     "prebuild": "rimraf dist",
     "build": "rollup -c",
@@ -24,12 +23,17 @@
   },
   "author": "Michael Ridgway <mcridgway@gmail.com>",
   "license": "BSD-3-Clause",
+  "dependencies": {
+    "react-is": "^16.7.0"
+  },
   "devDependencies": {
     "babel": "^6.23.0",
     "babel-cli": "^6.24.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-plugin-transform-react-jsx-source": "^6.22.0",
-    "babel-preset-env": "^1.6.1",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-register": "^6.24.1",
     "chai": "^4.0.1",
@@ -39,9 +43,11 @@
     "istanbul": "^0.4.5",
     "mocha": "^3.4.2",
     "pre-commit": "^1.0.7",
-    "react": "^15.0.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.7.0",
     "rimraf": "^2.6.2",
     "rollup": "^0.52.3",
+    "rollup-plugin-babel": "^3.0.4",
     "rollup-plugin-uglify": "^2.0.1"
   },
   "publishConfig": {
