@@ -10,12 +10,9 @@ import EntryElement from './entryElement.js';
 import firebase from 'firebase';
 import './entries.css';
 import {firebaseConfig} from 'Constants/defaultValues'
-var FeatureViewer = require("feature-viewer")
-import {
-  Label,
-  Input,
-  FormGroup,
-} from "reactstrap";
+import {Label, Input, FormGroup} from "reactstrap";
+import TextField from '@material-ui/core/TextField';
+
 
 
 firebase.initializeApp(firebaseConfig);
@@ -26,6 +23,7 @@ export default class extends Component {
     super(props);
     this.state = {
       entries    : [],
+      searchfield: '',
       loading    : true,
       start      : 1923192,
       end        : 2043192,
@@ -35,51 +33,23 @@ export default class extends Component {
 
   componentWillMount(){
     let entries = [];
-    firebase.firestore().collection("nORFs").where("chr", "==", this.state.chr).where("start", ">", 1923192).where("start", "<", 2043192).limit(200)
-                                            .get().then(function(data) {
-      data.forEach(function(doc) {
-          entries.push(doc.data());
-      })
-      }).catch(function(error) {
-        console.log("Error retrieving document:", error);
-    });
+    firebase.firestore().collection("nORFs").where("chr", "==", this.state.chr).where("start", ">", this.state.start).where("start", "<", this.state.end).limit(200)
+                        .get().then(function(data) {
+                                                      data.forEach(function(doc) {
+                                                          entries.push(doc.data());
+                                                      })
+                                                      }).catch(function(error) {
+                                                        console.log("Error retrieving document:", error);
+                                                    });
+
     this.setState({
       entries     : entries,
       loading     : false
     });
+
     console.log("componentWillMount complete");
   }
 
-  featureOverview(){
-
-      var ft = new FeatureViewer("0".repeat(this.state.end - this.state.start),
-                           '#peptideGraph',
-                            {
-                                showAxis: true,
-                                showSequence: false,
-                                brushActive: true, //zoom
-                                toolbar:false, //current zoom & mouse position
-                                bubbleHelp:false, 
-                                zoomMax:10 //define the maximum range of the zoom
-                            });
-
-    let objData = JSON.parse(JSON.stringify(this.state.entries).replace(/start/g, "x").replace(/end/g, "y"));
-    for(var i = 0; i < objData.length; i++){
-       objData[i]['y'] = objData[i]['y']-this.state.start > 10 ? objData[i]['y']-this.state.start : 10 ;
-       objData[i]['x'] = objData[i]['x']-this.state.start > 0 ? objData[i]['x']-this.state.start : 0 ;
-    }
-    console.log(objData);
-    ft.addFeature({
-      //data: [{"y":1255258,"x":1255223},{"y":7773506,"x":7773456},{"y":11087559,"x":11087515}],
-      data: objData,
-      name: "nORF",
-      className: "nORF", //can be used for styling
-      color: "#0F8292",
-      type: "rect" // ['rect', 'path', 'line']
-    });
-  
-
-  }
 
     componentDidUpdate(prevProps, prevState, snapshot){
       if(this.state.chr !== prevState.chr || this.state.start !== prevState.start || this.state.end !== prevState.end){
@@ -112,7 +82,36 @@ export default class extends Component {
     console.log(this.state.entries);
     console.log("searchHandle complete");
     this.setState({chr, start, end});
-    //this.featureOverview();
+
+    let entries = [];
+    firebase.firestore().collection("nORFs").where("chr", "==", this.state.chr).where("start", ">", this.state.start).where("start", "<", this.state.end).limit(200)
+                        .get().then(function(data) {
+                                                      data.forEach(function(doc) {
+                                                          entries.push(doc.data());
+                                                      })
+                                                      }).catch(function(error) {
+                                                        console.log("Error retrieving document:", error);
+                                                    });
+
+    this.setState({
+      entries     : entries,
+      loading     : false
+    });
+
+  }
+
+
+  searchHandle1() {
+    let search = this.state.searchfield;
+    let searchExp = /(chr|CHR)*\s*([0-9]{1,2}|X|Y|MT)\s*(-|:)?\s*(\d+)\s*(MB|M|K|)?\s*(-|:|)?\s*(\d+|)\s*(MB|M|K|)?/.exec(search);
+
+    let chr     = 'chr' + searchExp[2];
+    let start   = parseInt(searchExp[4]);
+    let end     = parseInt(searchExp[7]);
+    console.log(chr, start, end);
+    console.log(this.state.entries);
+    console.log("searchHandle complete");
+    this.setState({chr, start, end});
   }
 
   convertArrayOfObjectsToCSV(args) {
@@ -138,6 +137,8 @@ export default class extends Component {
       return result;
   }
 
+
+
   exportEntries() {
           let data, filename, link;
           let csv = this.convertArrayOfObjectsToCSV({
@@ -156,6 +157,9 @@ export default class extends Component {
   }
 
   render() {
+
+  
+
     let entryStore = this.state.entries && this.state.entries.map((entry, i) =>
                         <EntryElement
                           key = {'entry' + i}
@@ -173,29 +177,34 @@ export default class extends Component {
 
     const loader = this.state.loading ? <p> loading </p> : entryStore;
 
+  
+
+
     return (
       <Fragment>
         <Row>
         <Colxx xxs="8" className="middlesSearch">
-                <p> Beta Mode: Currently only chr1 is available! Please click twice on search due to a minor bug.</p>
         <div>
         <Row>
           <Colxx xxs="5" >
-            <FormGroup row>
-              <Input
-                type="text"
-                name="searchInput"
+             <TextField
+                style={{ margin: 8 }}
+                placeholder="Chr1:1-1239129"
+                fullWidth
+                margin="normal"
                 id="searchInput"
-                placeholder="chr1:54042-129391"
-                onKeyPress={(e) => {(e.key === 'Enter' ? this.searchHandle() : null)}}
-                //onChange={(evt) => {this.handleTagChange(evt.target.value);}}             
+                onChange={(event) => this.setState({searchfield : event.target.value})}
+                onKeyPress={(e) => {(e.key === 'Enter' ? this.searchHandle1() : null)}}
+                InputLabelProps={{
+                  shrink: true,
+                }}
               />
-            </FormGroup>
+
 
           </Colxx>
 
           <Colxx xxs="3">
-               <Button style={{marginRight: "5px"}} color="info" className="default mb-2" onClick={() => {this.searchHandle();}} >
+               <Button style={{marginRight: "5px"}} color="info" className="default mb-2" onClick={() => {this.searchHandle1();}} >
                   search
               </Button>
               {this.state.entries.length != 0 ? 
